refactor(tests): address PR feedback for logger tests (#8816)
